moved rate Parse and middleware callback functions to separate file
