+++ x/ratelimit/keeper/rate_limit.go
@@ -4,8 +4,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	transfertypes "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
-	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
 
 	"github.com/Stride-Labs/stride/v4/x/ratelimit/types"
 )
@@ -19,89 +17,6 @@ func GetRateLimitItemKey(denom string, channelId string) []byte {
 // is the total supply of the given denom
 func (k Keeper) GetChannelValue(ctx sdk.Context, denom string) sdk.Int {
 	return k.bankKeeper.GetSupply(ctx, denom).Amount
-}
-
-// Parse the denom from the Send Packet that will be used by the rate limit module
-// The denom that the rate limiter will use for a SEND packet depends on whether
-//    it was a NATIVE token (e.g. ustrd, stuatom, etc.) or NON-NATIVE token (e.g. ibc/...)...
-//
-// We can identify if the token is native or not by parsing the trace denom from the packet
-// If the token is NATIVE, it will not have a prefix (e.g. ustrd),
-//    and if it is NON-NATIVE, it will have a prefix (e.g. transfer/channel-2/uosmo)
-//
-// For NATIVE denoms, return as is (e.g. ustrd)
-// For NON-NATIVE denoms, take the ibc hash (e.g. hash "transfer/channel-2/usoms" into "ibc/...")
-func ParseDenomFromSendPacket(packet transfertypes.FungibleTokenPacketData) (denom string) {
-	// Determine the denom by looking at the denom trace path
-	denomTrace := transfertypes.ParseDenomTrace(packet.Denom)
-
-	// Native assets will have an empty trace path and can be returned as is
-	if denomTrace.Path == "" {
-		denom = packet.Denom
-	} else {
-		// Non-native assets should be hashed
-		denom = denomTrace.IBCDenom()
-	}
-
-	return denom
-}
-
-// Parse the denom from the Recv Packet that will be used by the rate limit module
-// The denom that the rate limiter will use for a RECIEVE packet depends on whether it was a source or sink
-// 		Source: The packet's is being recieved by a chain it was just sent from (i.e. the token has gone back and forth)
-//              (e.g. strd is sent -> to osmosis -> and then back to stride)
-//      Sink:   The packet's is being recieved by a chain that either created it or previous recieved it from somewhere else
-//              (e.g. atom is sent -> to stride) (e.g.2. atom is sent -> to osmosis -> which is then sent to stride)
-//
-//      If the chain is acting as a SINK:
-//      	We add on the Stride port and channel and hash it
-//          Ex1: uosmo sent from Osmosis to Stride
-//              Packet Denom:   uosmo
-//               -> Add Prefix: transfer/channel-X/uosmo
-//               -> Hash:       ibc/...
-//
-//          Ex2: ujuno sent from Osmosis to Stride
-//              PacketDenom:    transfer/channel-Y/ujuno  (channel-Y is the Juno <> Osmosis channel)
-//               -> Add Prefix: transfer/channel-X/transfer/channel-Y/ujuno
-//               -> Hash:       ibc/...
-//
-//      If the chain is acting as a SOURCE:
-//      	First, remove the prefix. Then if there is still a denom trace, hash it
-//          Ex1: ustrd sent back to Stride from Osmosis
-//              Packet Denom:      transfer/channel-X/ustrd
-//               -> Remove Prefix: ustrd
-//               -> Leave as is:   ustrd
-//
-//			Ex2: juno was sent to Stride, then to Osmosis, then back to Stride
-//              Packet Denom:      transfer/channel-X/transfer/channel-Z/ujuno
-//               -> Remove Prefix: transfer/channel-Z/ujuno
-//               -> Hash:          ibc/...
-func ParseDenomFromRecvPacket(packet channeltypes.Packet, packetData transfertypes.FungibleTokenPacketData) (denom string) {
-	// To determine the denom, first check whether Stride is acting as source
-	if transfertypes.ReceiverChainIsSource(packet.GetSourcePort(), packet.GetSourceChannel(), packetData.Denom) {
-		// Remove the source prefix (e.g. transfer/channel-X/transfer/channel-Z/ujuno -> transfer/channel-Z/ujuno)
-		sourcePrefix := transfertypes.GetDenomPrefix(packet.GetSourcePort(), packet.GetSourceChannel())
-		unprefixedDenom := packetData.Denom[len(sourcePrefix):]
-
-		// Native assets will have an empty trace path and can be returned as is
-		denomTrace := transfertypes.ParseDenomTrace(unprefixedDenom)
-		if denomTrace.Path == "" {
-			denom = unprefixedDenom
-		} else {
-			// Non-native assets should be hashed
-			denom = denomTrace.IBCDenom()
-		}
-	} else {
-		// Prefix the destination channel - this will contain the trailing slash (e.g. transfer/channel-X/)
-		destinationPrefix := transfertypes.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel())
-		prefixedDenom := destinationPrefix + packetData.Denom
-
-		// Hash the denom trace
-		denomTrace := transfertypes.ParseDenomTrace(prefixedDenom)
-		denom = denomTrace.IBCDenom()
-	}
-
-	return denom
 }
 
 // Checks whether the given packet will exceed the rate limit
